feat/SSAD-224/added-delete-method-into-controllers
